:sparkles: feat: incrementos na tela de not found.
